requirements.txt
anthropic==0.28.0
pydantic==2.6.4
python-dotenv==1.0.0
groq==0.9.0
google-generativeai==0.5.0
requests==2.31.0
structlog==24.1.0

# Optional hardware/vision backends
mujoco==3.1.5
opencv-python==4.9.0.80
ultralytics==8.1.34

# Testing
pytest==7.4.4
pytest-cov==4.1.0

